search: search for products feature done
