+++ OganiCommerce/search/urls.py
@@ -1,0 +1,10 @@
+from django.urls import path
+from . import views
+
+app_name = 'search'
+
+urlpatterns = [
+    path('', views.product_search, name='product_search'),
+
+
+]
