+++ OganiCommerce/store/urls.py
@@ -8,6 +8,4 @@ urlpatterns = [
     path('<slug:category_slug>/', views.product_list,
          name='product_list_by_category'),
     path('<int:id>/<slug:slug>/', views.product_detail, name='product_detail'),
-    path('search/<str:query>/', views.product_search, name='product_search'),
-
 ]
